src/main/java/models/Guitar.java
package models;

public class Guitar {
    private int id;
    private String name;
    private String brand;
    private String model;
    private int qty;
    private float price;



    public Guitar(int id, String name, String brand, String model, int qty, float price) {

        this.id = id;
        this.name = name;
        this.brand = brand;
        this.model = model;
        this.qty = qty;
        this.price = price;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getBrand() {
        return brand;
    }

    public void setBrand(String brand) {
        this.brand = brand;
    }

    public String getModel() {
        return model;
    }

    public void setModel(String model) {
        this.model = model;
    }

    public int getQty() {
        return qty;
    }

    public void setQty(int qty) {
        this.qty = qty;
    }

    public float getPrice() {
        return price;
    }

    public void setPrice(float price) {
        this.price = price;
    }

}
